fix: replace deprecated scheme.Builder with GroupVersionSchemeBuilder
controller-runtime v0.24 deprecated scheme.Builder (SA1019). Introduce
GroupVersionSchemeBuilder in pkg/util/scheme using only apimachinery,
replicating the same Register/AddToScheme behaviour. Update all five
groupversion_info.go files to use it.
