+++ api/cloud-control/v1beta1/groupversion_info.go
@@ -21,7 +21,8 @@ package v1beta1
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"sigs.k8s.io/controller-runtime/pkg/scheme"
+
+	utilscheme "github.com/kyma-project/cloud-manager/pkg/util/scheme"
 )
 
 var (
@@ -29,7 +30,7 @@ var (
 	GroupVersion = schema.GroupVersion{Group: "cloud-control.kyma-project.io", Version: "v1beta1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
-	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
+	SchemeBuilder = &utilscheme.GroupVersionSchemeBuilder{GroupVersion: GroupVersion}
 
 	// AddToScheme adds the types in this group-version to the given scheme.
 	AddToScheme = SchemeBuilder.AddToScheme
