+++ pkg/external/infrastructuremanagerv1/groupversion_info.go
@@ -3,7 +3,7 @@ package infrastructuremanagerv1
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/kyma-project/cloud-manager/pkg/util"
+	utilscheme "github.com/kyma-project/cloud-manager/pkg/util/scheme"
 )
 
 var (
@@ -11,7 +11,7 @@ var (
 	GroupVersion = schema.GroupVersion{Group: "infrastructuremanager.kyma-project.io", Version: "v1"} //nolint:gochecknoglobals
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
-	SchemeBuilder = &util.GroupVersionSchemeBuilder{GroupVersion: GroupVersion} //nolint:gochecknoglobals
+	SchemeBuilder = &utilscheme.GroupVersionSchemeBuilder{GroupVersion: GroupVersion} //nolint:gochecknoglobals
 
 	// AddToScheme adds the types in this group-version to the given scheme.
 	AddToScheme = SchemeBuilder.AddToScheme //nolint:gochecknoglobals
