+++ pkg/external/registrycachev1beta1/groupversion_info.go
@@ -3,7 +3,7 @@ package registrycachev1beta1
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/kyma-project/cloud-manager/pkg/util"
+	utilscheme "github.com/kyma-project/cloud-manager/pkg/util/scheme"
 )
 
 var (
@@ -11,7 +11,7 @@ var (
 	GroupVersion = schema.GroupVersion{Group: "core.kyma-project.io", Version: "v1beta1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme.
-	SchemeBuilder = &util.GroupVersionSchemeBuilder{GroupVersion: GroupVersion}
+	SchemeBuilder = &utilscheme.GroupVersionSchemeBuilder{GroupVersion: GroupVersion}
 
 	// AddToScheme adds the types in this group-version to the given scheme.
 	AddToScheme = SchemeBuilder.AddToScheme
